requirements.txt
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
slack_sdk==3.26.0
pytz==2023.3
slack-bolt==1.18.0
xlsxwriter>=3.0.0
httpx==0.26.0
h2==4.1.0
PyYAML>=6.0